Bluetooth: Controller: HCI: Fixing the bug to store 12 bit IQ samples
While IQ samples are captured in 12 bit resolution as described in
the documents of nRF chipsets, 4 bits of the data was ommited by
the software when it was reported to other components by HCI lasyer.

Now it stores the IQ samples as they are captured at the origin in
a 16 bit signed integer.

+++ include/bluetooth/hci.h
@@ -2442,8 +2442,8 @@ struct bt_hci_evt_le_chan_sel_algo {
 
 #define BT_HCI_EVT_LE_CONNECTIONLESS_IQ_REPORT  0x15
 struct bt_hci_le_iq_sample {
-	int8_t i;
-	int8_t q;
+	int16_t i;
+	int16_t q;
 };
 
 struct bt_hci_evt_le_connectionless_iq_report {
